Add all previous minor stable versions to current minor fast and stable
Today there's much confusion and uncertainty about the upgrade process
due to multiple factors. This may lead some people to switch to the next
channel before they're able to upgrade to the next minor. If they were on
a version that's not in the next channel they may be presented no upgrade
paths even if application of an additional z-stream update would unlock
a minor upgrade.

Hopefully this shouldn't be necessary when we push 4.7 out because the CVO
now knows which additional channels contain the current version enablign the
console to only present those channels when appropriate discouraging people
from changing prematurely.